Just apply target='_blank' (close #3)
When submitting, editing or deleting a comment, the post page will be
opened in a new window. And comments in the feed page is not changed.

This is not so good experience, but it makes maintenance easier.

Note:
If a "Submit" button in a comment form is clicked, eventually
jQuery(form).submit() will be called. The entire process is handled by
Qiita's JavaScript.

If a link to delete a comment is clicked, the event is caught by
"Ruby on Rails unobtrusive scripting adapter for jQuery"
(https://github.com/rails/jquery-ujs), and it sends a request to the server.